requirements.txt
asgiref==3.2.7
astroid==2.4.1
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.1
Django==3.0.6
django-environ==0.4.5
django-templated-mail==1.1.1
djangorestframework==3.11.0
drf-renderer-xlsx==0.3.7
et-xmlfile==1.0.1
idna==2.9
isort==4.3.21
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
openpyxl==3.0.3
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
wrapt==1.12.1
